Don't save our parent origin before it has been validated (#8)
#7 - Tabs SDK incorrectly updates parentOrigin if it receives a message from any origin

Today, when we receive any message that contains an object as its payload we update our parent origin to the origin of that message regardless of whether it's in our list of valid origins or not. A number of UI frameworks (e.g. Angular) use message posting to defer code execution and these messages set our parent origin to the origin of the window we are loaded in. This breaks all API calls made through our SDK since we end up using an incorrect target origin when posting messages to our parent window.

The fix for this is very straightforward - when processing an incoming message we need to delay updating our parent origin until we have validated the message's origin against our list of valid origins. I also added a unit test case to prevent this logic from regressing in the future.