Fix desktop client auth (#22)
This change addresses a couple of issues that were preventing the popup authentication flow in v0.3 of the SDK from working correctly in the Microsoft Teams desktop client. It also adds a unit test case that exercises the end-to-end authentication flow in the desktop client in order to catch issues like these in the future.

I verified the end-to-end authentication flow in the desktop client and all supported browsers (i.e. Chrome, Edge, Firefox, and IE).