Added ios and android hostClientTypes to authenticate() whitelist (#162)
**Description**
`microsoftTeams.authenticate()` is currently whitelisted to allow only the `desktop` client to pass the event to the Teams Shell. For other clients (web, ios, android), it triggers `window.open()`. Mobile will now support `microsoftTeams.authenticate()` and hence will need the function to propagate to the Host.

**Fix**
Added `ios` and `android` clients to the whitelist.

Fixes #163