Create utility function for constructing a chat deep link URL (#1063)
* parent a98e58484e28471417f3425abbe41d466948bf9a
author AE <36546967+AE-MS@users.noreply.github.com> 1649717360 -0700
committer AE <aaroner@microsoft.com> 1649717808 -0700

Create utility function for constructing a chat deep link URL

This utility function will be used to create a deep link URL that will
be sent to the host if it is determined that the host is "legacy teams"
and would not understand the chat capability messages.
This change does not actually execute the deep link (that will be in
the next change), it only constructs the deep link. It also includes
unit tests to validate the construction.