Switched from dynamic import of LiveShareClient to a global window var
The dynamic import approach to finding the LiveShareClient doesn't work when you get several layers of web packing going on. The `import()` function stops being able to find the code.

I had to switch to a simpler approach which just sets a global window variable that points at the the loaded LiveShareClient class. A GUID is used to avoid polluting the global namespace.