Stevenic/live share fix (#1381)
* Switched from dynamic import of LiveShareClient to a global window var

The dynamic import approach to finding the LiveShareClient doesn't work when you get several layers of web packing going on. The `import()` function stops being able to find the code.

I had to switch to a simpler approach which just sets a global window variable that points at the the loaded LiveShareClient class. A GUID is used to avoid polluting the global namespace.

* Change files

* Updated fluid references to match live-share

* Removed need for "import type" of fluid types

A small subset of types are directly defined in the liveShare namespace now.

* Adding in files that were removed by reversion of original live share PR

* Updated changefile text

Fix capitalization of LiveShareClient and wrapped in `` because it is a reference to something in code

* Removed liveShare namespace and exposed LiveShareHost class publicly from SDK

* Added LiveShareHost.create() method

* Marked LiveShareHost methods as hidden

* Added @beta tag to LiveShareHost methods

Co-authored-by: Trevor Harris <trharris@microsoft.com>