Making our use of the @beta tag consistent (#1411)
* Making our use of the @beta tag consistent

Our use of the @beta tag in this repo was not consistent for a couple of reasons. We initially thought that the @beta tag would cascade down and apply to things contained within something marked with @beta, but it does not.

After this change, I believe we now correctly have applied the beta tag everywhere. This means:

- If a namespace has a @beta tag on it, all children also have the @beta tag
- We no longer mark certain namespaces with @beta because we feel that making breaking changes there would now require bumping the major version of the package
- @beta tags indicate to readers of documentation that the item in question could change shape or disappear at any time and we will not bump the major version if something like that happens. When we feel that new things are sufficiently stable the @beta tag will be removed, at which point making breaking changes there would constitute a major version change.

* Change files

* Updated comment in dialog.ts