Make sure isSupported functions throw if checked before initialization is complete (#1430)
* Making supports empty by default

These values will be overridden as part of initialization and set to whatever the host actually supports. The way we used to do this, any capability that had subcapabilities would be marked as supported in order to then list the subcapabilities as undefined. This is not ideal!

* Change files

* Updated how we assign the unitialized runtime object and added unit tests

Added a lot of unit tests to cover whether any capabilities return that they are supported before initialization has been finished

* Make private isSupported functions require app initialization

* Updated all isSupported checks and unit tests in public folder

isSupported now calls ensureInitialized

* Fixed names of isSupported unit tests in private folder

They now say that they throw instead of return false if called before initialization

* Update @microsoft-teams-js-cc4b3c12-cbf1-4531-a77a-fe0f5c58b446.json

* Removed unused import from pages.spec.ts

Also includes some linting updates to order of import statements in test files

* Removing a lerna log (I don't know why it was generated) and updating .gitignore

* Update uninitialized runtime

apiVersion in _unitializedRuntime now starts at -1 (an illegal value) and will be overwritten with a legal value when we hear from the host

* Update all unit tests to use const Error for unitialized library

* Standardize comments on isSupported functions

* Added unit tests to cover null handler cases in some pages functions

* Remove unused import

I thought this was needed to make the docs work correctly, but it isn't and if I keep it in it causes a build failure

* Removing commented out import

* Store error message as string instead of Error to not mess up callstacks

* Updating @throws comments

* Added missing unit test for logs

registerGetLogHandler should not throw if called before initialization with a null handler (something we do during unit test cleanup sometimes)