Adding unit tests for communication.initializeCommunication (#1519)
* Adding some unit tests for the communication layer

More to come! This is adding a set of new tests around initializeCommunication, documenting how it works and what side effects it has

* Removed comment

* Change files

* Adding a couple more unit tests

Also fixed issues with two previously added unit tests that were failing in the pipeline because postMessage wasn't defined on the fake window objects. Fixes this by setting them to a window mock that DOES define post messag.e

* Updated unit test names and added a comment

Also changed the name of a new helper function in utils