Removed complicated regular expression that was breaking our YAML file (#1718)
The regular expression I was trying to use to only check pushes to 1.13.* or later for CodeQL issues was causing CodeQL to not run properly. I'm not sure what *exactly* the problem was, but I don't think there's any reason for a regex that complicated so I simplified it greatly. This change means that we'll now run codeql on any pushes to 1.* including branches earlier than 1.13, but that really shouldn't ever happen so I propose we do this for now and we can do something more complicated in the future if needed