Extend timeout of authentication unit test by 100ms
This test has been a little flaky in the past, primarily because the path we are unit testing has a currently unavoidable 300ms wait in it (100ms from communication.waitForMessageQueue and 200ms from authentication.notifySuccess. There is probably a clever way to track shortening those delays when using unit tests specifically, and we do have that tracked on our backlog. For now, let's extend this timeout so we have more than 50ms of wiggleroom.