Trharris/surfacehubauth (#1876)
* Added surfacehub to the list of host client types that send authenticate requests to the host instead of opening a window

* Change files

* auth unit test fixes and adding surfacehub to list of host types that ask host to do auth for them

* Extend timeout of authentication unit test by 100ms (#1875)

This test has been a little flaky in the past, primarily because the path we are unit testing has a currently unavoidable 300ms wait in it (100ms from communication.waitForMessageQueue and 200ms from authentication.notifySuccess. There is probably a clever way to track shortening those delays when using unit tests specifically, and we do have that tracked on our backlog. For now, let's extend this timeout so we have more than 50ms of wiggleroom.

* One unit test failure left

* Remove accidental handler change

* Fixed setInterval bug in authentication unit tests

* Update authentication.spec.ts

---------

Co-authored-by: AE ( ͡ಠ ʖ̯ ͡ಠ) <36546967+AE-MS@users.noreply.github.com>