Assume numberTypes have default min and max of 0
mustache doesn't have a great way to tell 0 apart from false as far as I can tell, and I don't want people to have to put numbers in as "0". As far as I can see, the best way to fix this is to make min and max numbers required (which is good practice anyways).