Add promise response to notify success. (#2589)
* Add promise response to notify success.

The notify success function, previously, was a fire and forget
function which hinder the possibility of the app knowing when
it finished in the host. This change modifies this and adds a response
to the function that indicates if the function finished.

Additionally, as we can't assure the host has already implemented the
answer back mechanism for this function a runtime property was added
that will determine if the host supports it or not. In this manner,
the `notifySuccess` function now needs the runtime to be populated
before even passing it to the communication layer.